+++ CHANGELOG.md
@@ -600,6 +600,7 @@
 
 - Bump Ruby and Rails versions
 - Update Rails configuration files with v6.1 settings
+- Add tests to guard against missing Activity field translations
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-45...HEAD
 [release-45]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-44...release-45